refactor(backend): move schema files to graphql folder
